+++ ByteCodeGenJobRef_.cpp
@@ -123,7 +123,7 @@ bool ByteCodeGenJob::emitStructDeRef(ByteCodeGenContext* context)
     if (typeInfo->kind == TypeInfoKind::Interface && (node->flags & (AST_FROM_UFCS | AST_TO_UFCS)))
     {
         if (node->flags & AST_FROM_UFCS) // Get the ITable pointer
-            emitInstruction(context, ByteCodeOp::DeRefPointer, node->resultRegisterRC, node->resultRegisterRC)->c.u32 = sizeof(void*);
+            emitInstruction(context, ByteCodeOp::DeRefPointer, node->resultRegisterRC, node->resultRegisterRC)->c.u64 = sizeof(void*);
         else if (node->flags & AST_TO_UFCS) // Get the structure pointer
             emitInstruction(context, ByteCodeOp::DeRefPointer, node->resultRegisterRC, node->resultRegisterRC);
         return true;
@@ -133,7 +133,7 @@ bool ByteCodeGenJob::emitStructDeRef(ByteCodeGenContext* context)
     {
         emitInstruction(context, ByteCodeOp::DeRefPointer, node->resultRegisterRC, node->resultRegisterRC);
         if (node->flags & AST_FROM_UFCS) // Get the ITable pointer
-            emitInstruction(context, ByteCodeOp::DeRefPointer, node->resultRegisterRC, node->resultRegisterRC)->c.u32 = sizeof(void*);
+            emitInstruction(context, ByteCodeOp::DeRefPointer, node->resultRegisterRC, node->resultRegisterRC)->c.u64 = sizeof(void*);
         else if (node->flags & AST_TO_UFCS) // Get the structure pointer
             emitInstruction(context, ByteCodeOp::DeRefPointer, node->resultRegisterRC, node->resultRegisterRC);
         return true;
